+++ GestorHotelero/src/app/components/admin-HOTEL/event/event.component.ts
@@ -97,6 +97,17 @@ export class EventComponent implements OnInit
   }
 
   deleteEvent(id: string) {
+    Swal.fire({
+      title: 'Are you sure to remove the event?',
+      text: 'This action cannot be reversed!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#aaa',
+      confirmButtonText: 'Yes, I want to delete it',
+      cancelButtonText: 'Cancell',
+    }).then((result) => {
+      if (result.isConfirmed) {
         this.eventRest.deleteEvent(this.idHotel, id).subscribe({
           next: (res: any) => {
             Swal.fire({
@@ -108,14 +119,17 @@ export class EventComponent implements OnInit
             });
             this.getEvents();
           },
-          error: (err) => Swal.fire({
+          error: (err) => { Swal.fire({
             title: err.error.message,
             icon: 'error',
             position: 'center',
             timer: 3000
-          })
-        })
-      }
+          });
+        },
+      });
+    }
+  });
+}
 
       updateEvent(){
         this.eventGetId.hotel= undefined;
